# Improve OpenType Layout performance