--- a/src/ot/layout.rs
+++ b/src/ot/layout.rs
@@ -40,9 +40,6 @@ impl<T> IndexMut<TableIndex> for [T] {
 
 /// A lookup-based layout table (GSUB or GPOS).
 pub trait LayoutTable {
-    /// The tag of this table.
-    const TAG: Tag;
-
     /// The index of this table.
     const INDEX: TableIndex;
 
@@ -53,7 +50,7 @@ pub trait LayoutTable {
     type Lookup: LayoutLookup;
 
     /// Get the lookup at the specified index.
-    fn get_lookup(&self, index: LookupIndex) -> Option<Self::Lookup>;
+    fn get_lookup(&self, index: LookupIndex) -> Option<&Self::Lookup>;
 }
 
 /// A lookup in a layout table.
@@ -163,7 +160,7 @@ pub fn apply_layout_table<T: LayoutTable>(
     plan: &ShapePlan,
     face: &Face,
     buffer: &mut Buffer,
-    table: Option<T>,
+    table: Option<&T>,
 ) {
     let mut ctx = ApplyContext::new(T::INDEX, face, buffer);
 
@@ -193,7 +190,7 @@ pub fn apply_layout_table<T: LayoutTable>(
     }
 }
 
-fn apply_string<T: LayoutTable>(ctx: &mut ApplyContext, lookup: T::Lookup) {
+fn apply_string<T: LayoutTable>(ctx: &mut ApplyContext, lookup: &T::Lookup) {
     if ctx.buffer.is_empty() || ctx.lookup_mask == 0 {
         return;
     }
@@ -225,7 +222,7 @@ fn apply_string<T: LayoutTable>(ctx: &mut ApplyContext, lookup: T::Lookup) {
     }
 }
 
-fn apply_forward(ctx: &mut ApplyContext, lookup: impl Apply) -> bool {
+fn apply_forward(ctx: &mut ApplyContext, lookup: &impl Apply) -> bool {
     let mut ret = false;
     while ctx.buffer.idx < ctx.buffer.len && ctx.buffer.successful {
         let cur = ctx.buffer.cur(0);
@@ -241,7 +238,7 @@ fn apply_forward(ctx: &mut ApplyContext, lookup: impl Apply) -> bool {
     ret
 }
 
-fn apply_backward(ctx: &mut ApplyContext, lookup: impl Apply) -> bool {
+fn apply_backward(ctx: &mut ApplyContext, lookup: &impl Apply) -> bool {
     let mut ret = false;
     loop {
         let cur = ctx.buffer.cur(0);