--- a/src/tables/gsub.rs
+++ b/src/tables/gsub.rs
@@ -3,19 +3,46 @@
 use super::gsubgpos::*;
 use super::*;
 
-#[derive(Clone, Copy, Debug)]
-pub struct SubstTable<'a>(pub SubstPosTable<'a>);
+#[derive(Clone, Debug)]
+pub struct SubstTable<'a> {
+    pub inner: SubstPosTable<'a>,
+    pub lookups: Vec<Option<SubstLookup<'a>>>,
+}
 
 impl<'a> SubstTable<'a> {
     pub fn parse(data: &'a [u8]) -> Option<Self> {
-        SubstPosTable::parse(data).map(Self)
+        let inner = SubstPosTable::parse(data)?;
+        let lookups = (0..inner.lookup_count())
+            .map(|i| inner.get_lookup(LookupIndex(i)).map(SubstLookup::parse))
+            .collect();
+
+        Some(Self { inner, lookups})
     }
 }
 
-#[derive(Clone, Copy, Debug)]
-pub struct SubstLookup<'a>(pub Lookup<'a>);
+#[derive(Clone, Debug)]
+pub struct SubstLookup<'a> {
+    pub subtables: Vec<SubstLookupSubtable<'a>>,
+    pub props: u32,
+    pub reverse: bool,
+}
 
-#[derive(Clone, Copy, Debug)]
+impl<'a> SubstLookup<'a> {
+    pub fn parse(lookup: Lookup<'a>) -> Self {
+        let subtables: Vec<_> = lookup
+            .subtables
+            .into_iter()
+            .flat_map(|data| SubstLookupSubtable::parse(data, lookup.kind))
+            .collect();
+
+        let props = lookup.props();
+        let reverse = subtables.iter().all(|subtable| subtable.is_reverse());
+
+        Self { subtables, props, reverse }
+    }
+}
+
+#[derive(Clone, Debug)]
 pub enum SubstLookupSubtable<'a> {
     Single(SingleSubst<'a>),
     Multiple(MultipleSubst<'a>),