@RazrFalcon 2020-12-04
Thanks! I've planed to work on this this weekends, but looks like I don't have to anymore.
And I've almost finished AAT. I have to port just one morx subtable and we're ready for a release.