@RazrFalcon 2020-12-04
As for the Latin performance, note that we have a very different TrueType parsing logic. Maybe in this case it makes a difference. Hard to tell.