@laurmaedje 2020-12-04
> And I've almost finished AAT. I have to port just one morx subtable and we're ready for a release.

 I'm looking forward to it :)